feat: use uncanto client for store tests
Our tests get pretty neat using the uncanto client to invoke things.

Also clears up some naming of the various principles so it's easier to figure out who is doing what to whom.

Aims to show off in tests what our normal flow should look like when folks delegate from a space to their agent first, and then invoke things on the space as the agent with the proof of delegation.

License: MIT
Signed-off-by: Oli Evans <oli@protocol.ai>