feat: carpark stack with bucket and eventbus (#36)
Carpark is one of the centre building blocks of our infrastructure. It
is a CAR store and several events are triggered from operations
happening to the bucket. Our HTTP API runs the UCANTO server and is an
abstraction layer to DynamoDB operations (and validating given CAR is
already in Carpark). With that said and aiming to keep SST Stacks tidy,
we should move the bucket from the API Stack into its own stack where we
can define all the resources and triggers. API Stack can access CARPARK
Stack to get Bucket reference (to get permissions to perform head
operation on it).

CARPARK stack includes:
- `carpark` bucket - previously on API Stack
- Event Bus responsible for handling S3 PUT Object requests and
forwarding them to appropriate targets
- Event Bus targets - for now focusing on minimal needs for MVP, so it
includes SQS Event to Elastic IPFS indexer and SQS Event to have CAR
file backup (and side index created)

resolves #13

Co-authored-by: Oli Evans <oli@protocol.ai>
Co-authored-by: Alan Shaw <alan.shaw@protocol.ai>