feat: replicator stack (#57)
Adds replicator stack by moving its content out of CARPARK on the
direction to #50

Implementation details:
- replicator-stack has 2 queues, one to replicate CARs and other to
replicate .idx files. At first I was attempting to do both in one single
queue. Ended up moving to 2 queues because I would need to have handlers
to identify origin bucket and have a hardcoded mapping where I would see
if it was carpark/satnav and with that select the appropriate CF
destination bucket.
- while my decision for 2 queues was based on the above, it might
actually be the best approach to not have queues throttling flow on
copies from independent buckets.
- also important noticing that despite 2 queues, there is only one
handler function code. The lambda code is the same, but it is injected
with destination BUCKET_NAME according to the queue it is handling.

Closes #54

Co-authored-by: Oli Evans <oli@protocol.ai>