refactor: move s3 url signing to car-store (#66)
- Add `createUploadUrl` to `car-store.js` as creating presigned s3 urls
is a job for the bucket abstraction and having a thing called `signer`
that wasn't a keypair for a DID for ucanto was confusing.
- Switch to the `@aws-sdk/s3-request-presigner` for creation of signed
s3 urls to simplify credential management. The existing s3 client is
reused.
- Adds `content-length` to the list of headers that form part of the url
signature.

fixes #38 

### Notes

When presigning a URL for s3, some headers are "hoisted" to the
URLSearchParams; they appear in the query string so you can hand the url
off as a string, and the params needed for auth are baked right in
there. S3 support unpacking those params from either the query _or_ from
http request headers. The query params are part of the signature, so
they can't be tampered with.

You can also "hoist" and sign any other params you want... it is
tempting to move `x-amz-checksum-sha256` to the query so a user wouldn't
need to provide it separately. However it appears that aws does not
support pulling that value out of the query, only from a request header.

It is unknown how s3 reacts to sending a content-length that doesn't
match the payload size, but there are tests for it in
https://github.com/web3-storage/sigv4/pull/7

License: MIT
Signed-off-by: Oli Evans <oli@protocol.ai>