fix: verify space registered on upload add (#111)
We previously added verification for `store/add`. We talked about doing
same for upload but we did not.

In the context of `store`, we did not add validation for `list` and
`remove` given it needed prior validations of a space being registered
for add operations. In `upload` we should do the same. Otherwise, there
is a vector where we can have our upload table written with uncontrolled
number of non existing rows.

Closes https://github.com/web3-storage/w3infra/issues/37