test: integration tests setup (#122)
This PR adds the basic setup for integration tests. It relies on
https://seed.run/docs/adding-a-post-deploy-phase.html to run the
integration tests once seed.run finishes its deploy. You can see output
in:
https://console.seed.run/dag-house/w3infra/activity/stages/pr122/builds/16/post-deploy
In case post deploy
[fails](https://console.seed.run/dag-house/w3infra/activity/stages/pr122/builds/15/post-deploy),
it is marked as failing in github CI too 👍🏼

After this approach is validated, we should in follow up PRs to the
actual interesting integration tests. Such as, using w3up client to
interact with upload API, and checking its dynamoDB, buckets, R2, etc to
guarantee everything happened as expect.

By default, not running `integration tests` with normal unit tests
script. Main reason for this is the way longer time needed with `sst
deploy` to run. We can have this in CI only, and be able to easily
replicate locally when we see errors in CI.

Alternatively, I tried to get tests to run in github action first. For
this, attempted to use
https://github.com/marketplace/actions/wait-for-github-deployment (as
well as other similar actions), but looks like we can't get these to
awake with seed.run releases (even though they in theory both use Github
Check API 🤷🏼‍♂️ ). After not being able to get this to work, I found out
the seed run post deploy trick. From one side, I would love to get these
to run in Github Actions, but I also feel this is ok while we have
deployment step also running in Seed.run. When (and if) we consider
moving this to github as well, we will be able to quite easily make the
integration tests also run in actions.

After merging:
- [ ] enable post deploy stage for all stages (except prod)