feat: add startCursor and endCursor
after working with dynamo a bit, I realized that adding
startCursor and endCursor to the response, combined with the new `pre`
caveat gives us fully stateless bi-directional paging!

I'd like to propose we deprecate `cursor` in favor of `startCursor` and
`endCursor`. even if we end up moving
to an implementation that doesn't support `startCursor`, `endCursor`
feels a bit clearer, and we don't have as tight space constraints in these
responses as we do in caveats that get included in UCANs