proposal: abstract pattern in upload and store service tests
We repeat the same pattern again and again in the upload and store service tests - creating the issuer, audience,
with and proofs and then constructing/invoking/executing capabilities
using them. After chatting with @vasco-santos last night I decided to throw
together a quick prototype of one way we might abstract this.

The downside of this is that these tests don't work quite as well as
usage examples of the capability code, which @vasco-santos said @olizilla
had previously advocated for. I'm down to try to find some better middle
ground here, so this PR is intended to be a starting
point for a conversation.

@vasco-santos @olizilla what do you think?