feat: ucan stream consumer system upload add count (#162)
Part of metrics work https://github.com/web3-storage/w3infra/issues/117

Adds system total metrics for `upload/add`. With these, we can track
number of uploads linked to a space. As we agreed, once we land receipts
we can revisit this to be associated with receipt to track uniqueness