feat: ucan stream endpoint and upload api with invocations and receipts (#172)
This PR includes two significant changes:
- `upload-api` ucanto server now follows same pattern as `access-api`
proposal from @Gozala's
https://github.com/web3-storage/w3protocol/pull/592
  - saves CAR file with UCAN invocations before executing them
  - sends one receipt per invocation when they are executed
- modifies `POST /ucan` endpoint to accept both receipts and CAR
invocations as discussed sync yesterday
- it identifies by ContentType wether is a CAR of invocations or a
Receipt (`processUcanLogRequest`)
  - for CAR of invocations, everything is quite similar as before:
    - parse CAR from bytes, persist CAR and send a view of it to Kinesis
- more than persisting the CARs, we also persist a map of invocation
CIDs to CARs where they live in same fashion as we do with DUDEWHERE
- kinesis now also receives a type property, which can either be
Invocation, or receipt like the actual endpoint
- NOTE: removed `proofs` as they can get quite big and we already talked
about getting rid of it before
  - for receipts:
- we parse the receipt CBOR, verify if given receipt is from an
Invocation that we previously got. If so, we can persist the Receipt
- Kinesis stream receives quite similar content to the invocation. Main
differences are that a receipt type is used, and we also include the
`inovcationCid`. In other words, we provide value from invocation so
that Kinesis consumers can know about `invocation` and `nb` from
invocation to filter out and account metrics

Other notes:
- `upload-api` hooks in the same functions for integrating with UCAN
Stream as the `/ucan` endpoint, avoiding the HTTP calls
- Storing receipts and invocations Bucket keys were decided yesterday. 
- Like https://github.com/web3-storage/w3protocol/pull/592, we aim to
get to ucanto part of the code that we currently inline to create the
receipts
- Will do a follow up PR for metric consumers to update to new world