feat: stop storing delegations in a bespoke bucket
Since we already store delegations inside invocations in the invocation log, we can simply store the invocation CID and delegation CID and then fetch and find the needed delegations when the are requested.

I'm not sure the implementation is correct, as I'm not super familiar with the distinction between invocations, workflows and receipts, but the types all check out!

Would love careful review on the new `findDelegationInInvocation` method I've added to `upload-api/tables/delegations.js`