feat: roundabout  (#205)
Adds roundabout :arrows_counterclockwise: stack to w3infra. It includes
two HTTP Endpoints:
- `GET /{carCid}` - Gets the content of a CAR file by its CID. Behind
the scenes, this request is redirected for a presigned URL created on
the spot
- `HEAD /{carCid}` - Gets header with `location` with presigned URL to
download CAR file
- Comes with a 3 days default expiration date (default chose given is
the actually limit imposed for a contract to be sealed in Spade), but
can be requested to be bigger via query param

This PR was added in this repo under the following decisions:
- we do not want to create more and more repos with SST projects,
specially for small context APIs
- we do not want to add this feature within the scope of
`up.web3.storage` and therefore a new API should be created

As drawback it is worth pointing out that creating more and more Stacks
can also bite us. There are account limits in the number of
Cloudformation Stacks and Roles we can have (we can also ask AWS to
grant us more).

CF docs for presigned URL
https://developers.cloudflare.com/r2/examples/aws/aws-sdk-js-v3/#generate-presigned-urls