feat: roundabout supports redirect to raw key with bucket as query param (#207)
There is also requirement to be able to create presigned URLs for
[dagcargo
bucket](https://dash.cloudflare.com/fffa4b4363a7e5250af8357087263b3a/r2/default/buckets/dagcargo).
It does not have same format as carpark.

This PR adds a `GET /key/{key}` endpoint to the API Gateway under
roundabout. It enables to pass a raw key of a bucket, rather than a CID
that is converted to the carpark key format. By default, it still picks
`carpark` bucket if no bucket is provided.

Integration tests were added so that we can actually test the
deployment, so that env vars and permissions are validated