feat: add date index to uploads table
The last piece of the blocking workflow that's not possible right now is being able to get a list of recent uploads. Add an index to the upload table to enable this.

I think this is the right thing to do for now while we figure out what kind of data warehousing setup we want, but we may want to consider removing this later.