feat: compute piece for uploaded cars (#228)
This is the first of a series of PRs to integrate w3up with w3filecoin.

This PR adds Filecoin Stack, where CARPARK bucket is hooked up with a
lambda to compute Piece CIDs via Event Bus events. A queue is in the
middle to keep the flow steady. Once a PR is computed, its details are
written into a DynamoDB table. This will have all the information users
might need to ask, in order to then ask Deal Tracker for information for
a given deal. In other words, it maps `piece`, `link` (Car CID),
`aggregate` and `inclusionProof` (TBD what it actually is, is it a CID
for a stored file in a bucket?).

Also integration tests were added.

In a follow up PR, a event will be hooked up from DynamoDB table insert
to trigger `segment/add` capability from Aggregator together with
writing a claim to equivalency claims.

TODO:
- [x] add
https://github.com/web3-storage/fr32-sha2-256-trunc254-padded-binary-tree-multihash/
published module

---------

Co-authored-by: Alan Shaw <alan.shaw@protocol.ai>