feat: rework revocation functions per new revocation design
The revocations table is designed to be batch-GET-queried by delegation CIDs, (which means the primary key MUST be just the delegation CID) but needs to accomodate multiple possible "revocation context CIDs" each with its own "cause CID".

If we could batch-get indices by primary key we could add fields for context and cause CIDs, but because BatchGetCommand only works on primary tables, not indices, we need to cram the context CID and the cause CIDs into a "set" field of :-separated contextCID:causeCID strings.

I'm not crazy about this design, but it works!