feat: integrate new storefront service and piece cid (#252)
Integrates new storefront service and events, together with the new
PieceCIDV2.

Note that:
- New Piece store DB
- the older one has outdated PieceCID versions, however we should be
able to re-use those to re-compute the new version and add them to new
table. Can be follow up migration when we want to also add backlog to
Filecoin
- As previously agreed, for MVP we allow agent to send us `piece/offer`
invocation, but we do not queue it as in
https://github.com/web3-storage/w3filecoin/blob/main/docs/w3filecoin.svg
- we rely on
https://github.com/web3-storage/w3up/blob/main/packages/filecoin-api/src/storefront/api.ts#L31
to skip submission queue on `piece/offer`. Instead, on bucket put, we
put it to the queue right away as we will not trust piece computations
at this point. However, this allows user sending correct piece to follow
receipt chain :)
- Unfortunately, when storing receipts we only store the out field,
which means we won't have the effects. Therefore, both receipt and task
stores will need to read from invocation bucket + workflow bucket.
- Irakli and I already talked about doing changes here once we land new
UCAN spec and have the tasks, so I suggest we deal with this as we do in
the PR for MVP