feat: filecoin metrics for ucan stream prometheus and updates metrics endpoint (#290)
Adds filecoin metrics for ucan stream in the same pattern we track other
metrics in https://up.web3.storage/metrics

We want to rely on UCAN Stream metrics table to in a "prometheus style"
increment metrics over time for totals. There are some metrics in
Firehose that we can rely on for ad-hoc queries and business queries
(like monthly usage). More than the price and efficiency in the long run
for totals in the Firehose, there are some limitations that are explicit
here:

- we use PieceCids to encode Piece size, which we will never be able to
handle in Athena level as we cannot decode the CID to get the size
- UCAN invocations (like aggregate/offer) may rely on attached blocks to
move attachments with the invocation, while only identifying the CID for
them. E.g in `aggregate/offer` we encode the very large list of pieces
to put into the aggregate in a CBOR block, and inline it in the CAR of
the invocation, while providing `pieces` as CID of the block in the nb
object.

We will need most measures I wanted to handle, except for tracking
execution time between the invocations. We need
https://github.com/web3-storage/w3up/issues/970 to get the power to
track execution time of a UCAN effect (e.g. time between receipt of
`aggregate/offer` and `aggregate/accept`