fix: roundabout should just redirect to 404 presigned url when no car in bucket instead of redundant check (#332)
@ribasushi did some profiling and roundabout is taking ~800ms just to
return the 302 status code (when requesting bytes by PieceCID). Probably
most of this time comes from waiting on the content claims response, but
also looking into what else Roundabout handler does, it may be a good
idea to just redirect to the presigned url without validating content is
in the bucket. We save one request to the bucket, and given we only use
roundabout in internal cases it should never be 404. But actually if it
is 404, might make more sense to then have presigned url response to be
404 than server to throw an error.