feat: upgrade storefront api with filecoin offer from bucket event instead of directly compute (#343)
This PR intends to
[upgrade](https://github.com/web3-storage/w3up/pull/1332) storefront
package. This enables `filecoin/offer` invocations lead to messages to
the submit queue, where message consumer will compute Piece for the
content. The main goal of this is to have the client to compute Filecoin
Pieces and offer them, instead of pieces being directly computed from
bucket event and written into the DB (skipping the submit validation).

Note that:
- given we will need to deal with client upgrades and old clients still
running for the time being, I opted in this PR to still keep a version
of the bucket event in place. However, instead of it running computation
of piece, it also triggers `filecoin/offer` so that we can keep old
clients running where `filecoin/offer` is not invoked. This of course
means that we may need to run Piece computation twice for the old path.
However, looking at our AWS bill, we paid 200 USD in
February for lambda execution of whole w3infra. This assumes then that we will
not incur a lot of additional cost while making migration much simpler
- datastore currently implements a composed store with S3+R2, which
allows us to move on an iterative approach where we can deploy first
step without client upgrade, while we can iterate on writes to anywhere
with the overall picture.