fix: upgrade filecoin api with content store to rely on roundabout (#360)
Today Storefront relies directly on a bucket (in S3) to read bytes to
derive PieceCID for validation. With blob protocol, we are moving to
write to R2 only 🙌🏼 therefore, we aim to make filecoin Storefront able
to read from anywhere, but also not increase the infra costs with this
change.

This PR changes Storefront to rely on Roundabout to read Blob bytes to
derive Piece CID from anywhere instead of `contentStore` being directly
hooked to a bucket. See
[usage](https://github.com/w3s-project/w3up/blob/main/packages/filecoin-api/src/storefront/events.js#L43)

Relying on Roundabout introduces a layer of indirection, but in current
implementation where this validation runs in AWS and content will likely
be read from CF, makes it cost optimal to read from Roundabout.

Closes https://github.com/w3s-project/w3up/issues/1349